backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
celery==5.3.0
redis==4.5.5
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1